+++ app/Http/Controllers/UserController.php
@@ -22,7 +22,6 @@ class UserController extends Controller
     public function __construct(protected UserRepositoryInterface $user)
     {
         parent::__construct();
-        $this->authorizeResource(User::class, 'user');
     }
 
     /**
@@ -30,6 +29,8 @@ class UserController extends Controller
      */
     public function index(Request $request): Response
     {
+        $this->authorize('read-user', User::class);
+
         $users = $this->user->getAll($request);
         $resource = UserResource::collection($users);
 
@@ -41,6 +42,8 @@ class UserController extends Controller
      */
     public function create(): Response
     {
+        $this->authorize('create-user', User::class);
+
         return InertiaResponse::content('Users/Create');
     }
 
@@ -49,6 +52,8 @@ class UserController extends Controller
      */
     public function store(StoreUserRequest $request): RedirectResponse
     {
+        $this->authorize('create-user', User::class);
+
         $user = $this->user->newUser($request->all());
         $user->assignRole($request->get('role'));
         if ($request->hasFile('avatar')) {
@@ -64,6 +69,8 @@ class UserController extends Controller
     public function show(int $id): Response
     {
         try {
+            $this->authorize('read-user', User::class);
+
             $user = $this->user->getById($id);
             $resource = new UserResource($user);
 
@@ -83,6 +90,8 @@ class UserController extends Controller
     public function edit(int $id): Response
     {
         try {
+            $this->authorize('update-user', User::class);
+
             $user = $this->user->getById($id);
             $resource = new UserResource($user);
 
@@ -102,6 +111,7 @@ class UserController extends Controller
     public function update(UpdateUserRequest $request, int $id): Response|RedirectResponse
     {
         try {
+            $this->authorize('update-user', User::class);
 
             $user = $this->user->updateUser($request->all(), $id);
 
@@ -132,6 +142,8 @@ class UserController extends Controller
     public function destroy(int $id): Response|RedirectResponse
     {
         try {
+            $this->authorize('delete-user', User::class);
+
             $this->user->deleteUser($id);
 
             return to_route('users.index')->with('success', 'User deleted successfully');
